Pub git2_hooks::create_hook only for feature "test-utils"
git2_hooks::create_hook is a function used in tests, which panics on common
errors. It can't really be moved into git2-testing, as it depends on
git2_hooks::hookspath::HookPaths.

In order to avoid accidental use, move the function into
git2_hooks::test_utils[_priv] and only publish it as git2_hooks::test_utils
if feature = "test-utils" is enabled.